Start to create a basic MergerFeatureTest for the merging issue
